+++ src/Entity.hpp
@@ -14,13 +14,13 @@
 
 class Entity{
 private:
-    int x, y;
+    float x, y;
 public:
     ofPoint position;
     bool visible;
     ofImage texture;
     std::string imgPath;
-    Entity(int x, int y, bool visible, std::string imgPath);
+    Entity(float x, float y, bool visible, std::string imgPath);
     void display();
 };
